♻️ [1/3] dynamic scheduler structure refactor round (#3698)
