+++ services/director-v2/tests/unit/conftest.py
@@ -416,7 +416,7 @@ def mock_docker_api(mocker: MockerFixture) -> None:
         return_value=[],
     )
     mocker.patch(
-        "simcore_service_director_v2.modules.dynamic_sidecar.scheduler._task_utils.are_sidecar_and_proxy_services_present",
+        "simcore_service_director_v2.modules.dynamic_sidecar.scheduler._core.are_sidecar_and_proxy_services_present",
         autospec=True,
         return_value=True,
     )
